+++ R/mirt.R
@@ -239,7 +239,8 @@
 #'
 #' @aliases mirt
 #' @param data a \code{matrix} or \code{data.frame} that consists of
-#'   numerically ordered data, with missing data coded as \code{NA}
+#'   numerically ordered data, with missing data coded as \code{NA} (to convert from an ordered factor
+#'   \code{data.frame} see \code{\link{data.matrix}})
 #' @param model a string to be passed (or an object returned from) \code{\link{mirt.model}},
 #'   declaring how the IRT model is to be estimated (loadings, constraints, priors, etc).
 #'   For exploratory IRT models, a single numeric value indicating the number
@@ -265,11 +266,11 @@
 #'   The \code{'EM'} is generally effective with 1-3 factors, but methods such as the \code{'QMCEM'}
 #'   or \code{'MHRM'} should be used when the dimensions are 3 or more
 #' @param optimizer a character indicating which numerical optimizer to use. By default, the EM
-#'   algorithm will use the \code{'BFGS'} when there are no upper and lower bounds, and a
-#'   penelized version of the BFGS algorithm when there are.
+#'   algorithm will use the \code{'BFGS'} when there are no upper and lower bounds, and
+#'   \code{'L-BFGS-B'} when there are. Another good option which supports bound constraints is
+#'   the \code{'nlminb'}, which is more stable than the BFGS family of optimizers (though slightly slower).
 #'
-#'   Another good option which supports bound constraints is
-#'   the \code{'nlminb'}. Other options include the Newton-Raphson (\code{'NR'}),
+#'   Other options include the Newton-Raphson (\code{'NR'}),
 #'   which often will be more efficient than the \code{'BFGS'} but not as stable for more complex
 #'   IRT models (such as the nominal or nested logit models) and does not support
 #'   upper and lower bound constraints. As well, the \code{'Nelder-Mead'}, \code{'SANN'}, and \code{'L-BFGS-B'}
@@ -977,7 +978,7 @@
 #' }
 mirt <- function(data, model, itemtype = NULL, guess = 0, upper = 1, SE = FALSE,
                  covdata = NULL, formula = NULL, SE.type = 'crossprod', method = 'EM',
-                 optimizer = NULL, pars = NULL, constrain = NULL, parprior = NULL,
+                 optimizer = 'BFGS', pars = NULL, constrain = NULL, parprior = NULL,
                  calcNull = TRUE, draws = 5000, survey.weights = NULL,
                  quadpts = NULL, TOL = NULL, gpcm_mats = list(), grsm.block = NULL,
                  rsm.block = NULL, key = NULL,